admin panel and reseravation list pages added

+++ database/migrations/2021_05_17_040656_create_rules_table.php
@@ -16,7 +16,7 @@ class CreateRulesTable extends Migration
         Schema::create('rules', function (Blueprint $table) {
             $table->id();
             $table->string('rule');
-            $table->integer('value');
+            $table->double('value');
             $table->boolean('active');
             $table->timestamps();
         });
